Revert "optimized languageFetcher"
This reverts commit 3ee38f011493ba5e974fd77f81c870b22f32793f.
