+++ issues.go
@@ -17,7 +17,6 @@ type Issue struct {
 	Title     string
 	Date      time.Time
 	URL       string
-	RepoURL   string
 	Repo      Repo
 	Languages []string
 }
@@ -53,10 +52,6 @@ func fetchIssues(ctx context.Context, token string) ([]Issue, error) {
 	// main chan where workers send their results
 	ch := make(chan Issue)
 
-	// Setup and start a languageFetcher
-	lf := newLanguageFetcher()
-	go lf.start()
-
 	// errors is where workers will report failure. It has to have sufficient
 	// buffer space to prevent deadlocks because we only receive from it once
 	errors := make(chan error, len(labels))
@@ -70,7 +65,7 @@ func fetchIssues(ctx context.Context, token string) ([]Issue, error) {
 	wg.Add(len(labels))
 	for _, l := range labels {
 		go func(l string) {
-			if err := issueSearch(cCtx, lf, l, token, ch); err != nil {
+			if err := issueSearch(cCtx, l, token, ch); err != nil {
 				errors <- err
 			}
 			wg.Done()
@@ -96,12 +91,7 @@ func fetchIssues(ctx context.Context, token string) ([]Issue, error) {
 		// our results and send them up. If it was open just append the value.
 		case i, open := <-ch:
 			if !open {
-				// Check for errors and wait for language workers to finish.
-				if err := lf.wait(cancel); err != nil {
-					return nil, err
-				}
-				// Append repo languages to deduped issues.
-				return lf.appendLanguages(dedupe(issues)), nil
+				return dedupe(issues), nil
 			}
 			issues = append(issues, i)
 		}
@@ -126,7 +116,7 @@ func dedupe(in []Issue) []Issue {
 // into ch as they are found. An error is returned if we could not complete the
 // request or GitHub responds with anything but a 200. A ctx is provided so we
 // know if we need to quit early.
-func issueSearch(ctx context.Context, lf *languageFetcher, label, token string, ch chan<- Issue) error {
+func issueSearch(ctx context.Context, label, token string, ch chan<- Issue) error {
 	ctx.Done()
 
 	req, err := http.NewRequest("GET", "https://api.github.com/search/issues", nil)
@@ -180,17 +170,17 @@ func issueSearch(ctx context.Context, lf *languageFetcher, label, token string, 
 	}
 
 	for _, item := range data.Items {
-		lf.wg.Add(1)
-		go lf.repoLanguages(ctx, item.RepoURL, token)
+		lf := newLanguageFetcher()
+		languages, err := lf.repoLanguages(ctx, item.RepoURL, token)
 		if err != nil {
 			return err
 		}
 
 		issue := Issue{
-			Title:   item.Title,
-			Date:    item.CreatedAt,
-			URL:     item.URL,
-			RepoURL: item.RepoURL,
+			Title:     item.Title,
+			Date:      item.CreatedAt,
+			URL:       item.URL,
+			Languages: languages,
 		}
 
 		issue.Repo, err = repoFromURL(item.RepoURL)
@@ -209,4 +199,56 @@ func issueSearch(ctx context.Context, lf *languageFetcher, label, token string, 
 		}
 	}
 	return nil
+}
+
+type languageFetcher struct {
+	fetchedRepos map[string][]string
+}
+
+func newLanguageFetcher() *languageFetcher {
+	return &languageFetcher{
+		fetchedRepos: make(map[string][]string),
+	}
+}
+
+func (lf *languageFetcher) repoLanguages(ctx context.Context, repoURL, token string) ([]string, error) {
+	// Return cached languages if already fetched from repo.
+	if langs := lf.fetchedRepos[repoURL]; langs != nil {
+		return langs, nil
+	}
+
+	// If not cached, get languages from repo.
+	req, err := http.NewRequest("GET", fmt.Sprintf(repoURL+"/languages"), nil)
+	if err != nil {
+		return nil, errors.Wrap(err, "could not build request")
+	}
+
+	// Tell the request to use our context so we can cancel it in-flight if needed
+	req = req.WithContext(ctx)
+
+	// Use their access token so it counts against their rate limit
+	if token != "" {
+		req.Header.Add("Authorization", "token "+token)
+	}
+
+	resp, err := http.DefaultClient.Do(req)
+	if err != nil {
+		return nil, errors.Wrap(err, "could not execute request")
+	}
+	defer resp.Body.Close()
+
+	if resp.StatusCode != 200 {
+		return nil, errors.Wrapf(err, "status was %d, not 200", resp.StatusCode)
+	}
+	data := make(map[string]int)
+	if err := json.NewDecoder(resp.Body).Decode(&data); err != nil {
+		return nil, errors.Wrap(err, "could not decode json")
+	}
+
+	// Get top three languages
+	langs := top(3, data)
+
+	// Cache repo languages.
+	lf.fetchedRepos[repoURL] = langs
+	return langs, nil
 }
